corpus: bh -- float() around floor(); add PYC_CHANGES.md with the editing policy
Author's call: edit the corpus here, comment the source, and write the
policy down. This does all three.

THE EDIT. `Vec3.__init__` writes `self.d0 = 0.0` (float) and `xp[0] =
floor(...)` writes an int into the SAME object -- CPython 3's math.floor
returns int. One object holding float then int over its lifetime is
TEMPORAL, not per-object, so no contour split can separate it; the field's
static type is {int, float}, which has no unboxed representation. Boxing, a
semantic deviation, or a cast in the source are the only three answers, and
the goal (CPython semantics for programs that can be made monotonic) rules
out the first two.

Semantically a no-op, verified not assumed: every read goes through int(...)
(`int(xp[k]) & l`, `int(ic[k]) & l`), Node.IMAX is 2**30 and 0.0 <= xsc <
1.0, so floor(IMAX*xsc) is in [0, 2**30] and exact in a double. CPython
output is BYTE-IDENTICAL before and after. pyc at the default arm compiles,
runs, and matches CPython. Flag arm: ess 412 -> 366, css 1164 -> 1134,
Vec3 contours 20 -> 6 (measured on a scratch copy), route-4 mints 24 -> 7.

Six sites, one carrying the full rationale as a comment.

NOTE shedskin compiles this file unedited, and that is NOT evidence pyc is
missing something: shedskin's math.floor returns __ss_float
(lib/math/__init__.hpp:36) where CPython's returns int, so it never creates
the mix. Its clean typing here is bought by a semantic deviation.

THE POLICY, in shedskin_examples/PYC_CHANGES.md. A corpus edit is allowed
only when it is a VERIFIED CPython no-op, says what the code already meant,
beats any flag that would do instead without a semantic cost, carries its
reason as a comment at the edit, and is tabled in that file. Editing around
a pyc inference deficiency is not allowed -- the test is whether CPython
itself would have to box the value.

The table records all three changes from upstream: chess printBoard
(9abf39f9), chess `return False` (3cfdbd7d), and this one, plus the CRLF
conversion (34fc4fb5) so the diff against upstream is fully accounted for.

Deliberately a NEW file rather than the existing README.md, which is
upstream shedskin's program index -- this directory is a subtree and
editing upstream files invites merge conflicts. (I overwrote it first and
restored it.)

Also records the cost this policy carries: the CPython cache key covers
every corpus **/*.py, so touching one file invalidates all 77 CPython
results. They re-run cold and every rc=124 is re-taken alone -- measured
here as 20 timeouts and a ~12-minute sweep becoming over an hour.

Gates: make test, LLVM e2e, test_llvm, test_dparse, test_links all pass.
Corpus sweeps were still running at commit time (the cache rebuild above);
only bh can change, since the pyc binary is unchanged from the arm already
swept clean at 4a149bc0 and no other corpus source was touched.

bh still FAILS under the flag -- 11 `illegal: str` warnings and a runtime
abort. That is ifa/143 (three __slots__ string lists merging into the node
lists), untouched by this and independent of the numeric mix.

+++ shedskin_examples/bh/bh.py
@@ -348,13 +348,32 @@ class Body(Node):
         xp = Vec3()
 
         xsc = (self.pos[0] - tree.rmin[0]) / tree.rsize
-        xp[0] = floor(Node.IMAX * xsc)
+        # pyc: `float(...)` around `floor(...)`, six sites in this file.
+        #
+        # `Vec3.__init__` writes `self.d0 = 0.0`, so a fresh Vec3's fields
+        # are float; `floor()` returns an int in CPython 3, so this write
+        # puts an int into the SAME object. One object holding float then
+        # int over its lifetime is TEMPORAL, not per-object, so no contour
+        # split can separate it -- the field's static type is `{int, float}`
+        # and that has no unboxed representation.
+        #
+        # Semantically a no-op: every read of these fields goes through
+        # `int(...)` (`int(xp[k]) & l`, `int(ic[k]) & l`), `Node.IMAX` is
+        # 2**30 and `0.0 <= xsc < 1.0`, so `floor(IMAX*xsc)` lands in
+        # [0, 2**30] and is exact in a double. `int(float(floor(x)))` is
+        # `floor(x)`.
+        #
+        # See shedskin_examples/PYC_CHANGES.md for the policy. NOTE shedskin
+        # compiles this file unedited only because ITS `math.floor` returns
+        # a double (lib/math/__init__.hpp:36), which CPython's does not --
+        # so this is not a pyc deficiency shedskin lacks.
+        xp[0] = float(floor(Node.IMAX * xsc))
 
         xsc = (self.pos[1] - tree.rmin[1]) / tree.rsize
-        xp[1] = floor(Node.IMAX * xsc)
+        xp[1] = float(floor(Node.IMAX * xsc))
 
         xsc = (self.pos[2] - tree.rmin[2]) / tree.rsize
-        xp[2] = floor(Node.IMAX * xsc)
+        xp[2] = float(floor(Node.IMAX * xsc))
 
         i = 0
         for k in range(Vec3.NDIM):
@@ -579,19 +598,19 @@ class Tree:
 
         xsc = (vp[0] - self.rmin[0]) / self.rsize
         if 0.0 <= xsc and xsc < 1.0:
-            xp[0] = floor(Node.IMAX * xsc)
+            xp[0] = float(floor(Node.IMAX * xsc))
         else:
             return None
 
         xsc = (vp[1] - self.rmin[1]) / self.rsize
         if 0.0 <= xsc and xsc < 1.0:
-            xp[1] = floor(Node.IMAX * xsc)
+            xp[1] = float(floor(Node.IMAX * xsc))
         else:
             return None
 
         xsc = (vp[2] - self.rmin[2]) / self.rsize
         if 0.0 <= xsc and xsc < 1.0:
-            xp[2] = floor(Node.IMAX * xsc)
+            xp[2] = float(floor(Node.IMAX * xsc))
         else:
             return None
